+++ src/inputs/mod.rs
@@ -1,0 +1,3 @@
+pub mod menu;
+pub mod pause;
+pub mod ship;